move backend code into separate bundle
