+++ main/helpers/fork-backend.js
@@ -1,19 +1,20 @@
+import path from 'path';
 import { fork } from 'child_process';
 import { app } from 'electron';
 
 export default function forkBackend() {
-  return new Promise(resolve => {
+  return new Promise((resolve) => {
     let workerProcess;
-    const handleReady = message => {
+    const handleReady = (message) => {
       if (message === 'ready') {
         workerProcess.removeListener('message', handleReady);
         resolve(workerProcess);
       }
     };
 
-    workerProcess = fork(__filename, ['--backend']);
+    workerProcess = fork(path.join(__dirname, 'backend.js'));
     workerProcess.on('message', handleReady);
-    workerProcess.on('exit', code => {
+    workerProcess.on('exit', (code) => {
       if (code !== 0 && code !== null) {
         throw new Error(
           `Backend thread unexpectedly exited with code ${code}.`
